Skill cards display fixed
